atualizacao e organizacao do projeto, para um melhor entendimento e estrutura
